fix(cli): preserve isolated-home path boundary
Classify ASTRID_HOME before PATH setup can create its binary directory or
edit a shell profile, and move the PATH contract into a private module so
the idempotency rationale remains explicit. Exercise the real boundary in
a disposable subprocess.
